split etb interface
