+++ firmware/controllers/actuators/electronic_throttle.cpp
@@ -75,7 +75,8 @@
 
 #if EFI_ELECTRONIC_THROTTLE_BODY
 
-#include "electronic_throttle.h"
+#include "electronic_throttle_impl.h"
+#include "engine.h"
 #include "tps.h"
 #include "sensor.h"
 #include "dc_motor.h"
